+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 
+- Update `code_text` to support Dark Mode with `glimmer_dark` theme, which is applied by default when app is run with OS in dark mode
 - Fix issue with `code_text` showing gibberish on Linux when Courier & Consolas fonts are not available. It will attempt to use the first Monospace font it finds now or otherwise stick to the default font.
 - Refactor Hello, Custom Widget! with on_widget_disposed thread cleanup code
 - Refactor Hello, Custom Shell! to support keyboard display of emails by hitting ENTER/SPACE after selecting a table row